Update window control bar as a custom control Update crud buttons style Admin UIs - Update search bar
