+++ ThesisManagement/App.xaml
@@ -2,7 +2,7 @@
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:local="clr-namespace:ThesisManagement"
-             StartupUri="Views/Professor/ProfessorMainView.xaml">
+             StartupUri="Views/Admin/AdminMainView.xaml">
     
     <Application.Resources>
         <ResourceDictionary>
