+++ ThesisManagement/CustomControls/WindowControlBar.xaml
@@ -1,0 +1,31 @@
+﻿<UserControl x:Class="ThesisManagement.CustomControls.WindowControlBar"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:ThesisManagement.CustomControls"
+             xmlns:fa="http://schemas.awesome.incremented/wpf/xaml/fontawesome.sharp"
+             mc:Ignorable="d" 
+             Background="White"
+             d:DesignHeight="25" d:DesignWidth="800">
+    <StackPanel Orientation="Horizontal"
+                FlowDirection="RightToLeft">
+        <Button x:Name="btnClose"
+                Click="btnClose_Click"
+                Style="{StaticResource windowControlBtn}">
+             <fa:IconImage Icon="X" Style="{StaticResource windowControlIcon}" />
+        </Button>
+
+        <Button x:Name="btnMaximize"
+                Click="btnMaximize_Click"
+                Style="{StaticResource windowControlBtn}">
+            <fa:IconImage Icon="Square" Style="{StaticResource windowControlIcon}" />
+        </Button>
+
+        <Button x:Name="btnMinimize"
+                Click="btnMinimize_Click"
+                Style="{StaticResource windowControlBtn}">
+            <fa:IconImage Icon="WindowMinimize" Style="{StaticResource windowControlIcon}" />
+        </Button>
+    </StackPanel>
+</UserControl>
