+++ ThesisManagement/CustomControls/WindowControlBar.xaml.cs
@@ -1,0 +1,47 @@
+﻿using System.Runtime.InteropServices;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Input;
+using System.Windows.Interop;
+
+namespace ThesisManagement.CustomControls
+{
+    public partial class WindowControlBar : UserControl
+    {
+        [DllImport("user32.dll")]
+        public static extern IntPtr SendMessage(IntPtr hWnd, int wMsg, int wParam, int lParam);
+
+        public WindowControlBar()
+        {
+            InitializeComponent();
+            // Add a handler for the MouseLeftButtonDown event with handledEventsToo set to true
+            this.AddHandler(UIElement.MouseLeftButtonDownEvent, new MouseButtonEventHandler(OnMouseLeftButtonDown), true);
+        }
+
+        private void btnMinimize_Click(object sender, RoutedEventArgs e)
+        {
+            Window.GetWindow(this).WindowState = WindowState.Minimized;
+        }
+
+        private void btnMaximize_Click(object sender, RoutedEventArgs e)
+        {
+            if (Window.GetWindow(this).WindowState == WindowState.Normal)
+            {
+                Window.GetWindow(this).WindowState = WindowState.Maximized;
+            }
+            else
+                Window.GetWindow(this).WindowState = WindowState.Normal;
+        }
+
+        private void btnClose_Click(object sender, RoutedEventArgs e)
+        {
+            Application.Current.Shutdown();
+        }
+
+        private void OnMouseLeftButtonDown(object sender, MouseButtonEventArgs e)
+        {
+            WindowInteropHelper helper = new WindowInteropHelper(Window.GetWindow(this));
+            SendMessage(helper.Handle, 161, 2, 0);
+        }
+    }
+}
