+++ ThesisManagement/Resources/Styles/Buttons.xaml
@@ -69,45 +69,10 @@
         </Style.Triggers>
     </Style>
 
-    <!--Control bar button-->
-    <Style x:Key = "controlButton" TargetType = "Button" >
-        <!--Normal button style-->
-        <Setter Property = "Width" Value = "35" />
-        <Setter Property = "Background" Value = "Transparent" />
-        <Setter Property = "Foreground" Value = "{StaticResource plainTextColor}" />
-        <Setter Property = "BorderBrush" Value = "Transparent" />
-        <Setter Property = "Template" >
-            <Setter.Value >
-                <ControlTemplate TargetType = "Button" >
-                    <Border Background = "{TemplateBinding Background}">
-                        <ContentPresenter HorizontalAlignment = "Center" VerticalAlignment = "Center" />
-                    </Border>
-                </ControlTemplate>
-            </Setter.Value>
-        </Setter>
-        <Style.Triggers>
-            <!--Button Style on hover-->
-            <Trigger Property = "IsMouseOver" Value = "True" >
-                <Setter Property = "Background" Value = "{StaticResource panelActiveColor}" />
-                <Setter Property = "Foreground" Value = "{Binding Path=Tag, RelativeSource={RelativeSource Self}}" />
-            </Trigger>
-        </Style.Triggers>
-    </Style>
-
-    <!--Control bar button icon-->
-    <Style x:Key = "controlButtonIcon" TargetType = "fa:IconImage" >
-        <Setter Property = "Foreground" Value = "{Binding Path=Tag, RelativeSource={RelativeSource AncestorType=Button}}" />
-        <Setter Property = "Width" Value = "12" />
-        <Setter Property = "Height" Value = "12" />
-    </Style>
-
-    <!--Icon user button-->
-    <Style x:Key = "iconButton" TargetType = "Button" >
-        <!--Normal button style-->
+    <Style x:Key = "profileBtn" TargetType = "Button" >
         <Setter Property="Cursor" Value="Hand"/>
         <Setter Property = "Width" Value = "25" />
         <Setter Property = "Background" Value = "Transparent" />
-        <Setter Property = "Foreground" Value = "{StaticResource plainTextColor}" />
         <Setter Property = "Template" >
             <Setter.Value >
                 <ControlTemplate TargetType = "Button" >
@@ -118,25 +83,20 @@
             </Setter.Value>
         </Setter>
         <Style.Triggers>
-            <!--Button Style on hover-->
             <Trigger Property = "IsMouseOver" Value = "True" >
                 <Setter Property = "Foreground" Value = "{Binding Path=Tag, RelativeSource={RelativeSource Self}}" />
             </Trigger>
         </Style.Triggers>
     </Style>
 
-    <!--Icon user button icon-->
-    <Style x:Key = "iconButtonIcon" TargetType = "fa:IconImage" >
-        <Setter Property = "Foreground" Value = "{Binding Path=Tag, RelativeSource={RelativeSource AncestorType=Button}}" />
-        <Setter Property = "Width" Value = "16" />
-        <Setter Property = "Height" Value = "16" />
+    <Style x:Key = "profileBtnIcon" TargetType = "fa:IconImage" >
+        <Setter Property = "Width" Value = "10" />
+        <Setter Property = "Height" Value = "10" />
     </Style>
     
-    <!--Cruds Button-->
     <Style x:Key="btnCreate" TargetType="Button">
         <Setter Property="Background" Value="{StaticResource createNomal}"/>
         <Setter Property="FontSize" Value="12"/>
-        <Setter Property="FontFamily" Value="Montserrat"/>
         <Setter Property="Cursor" Value="Hand"/>
         <Setter Property="BorderThickness" Value="0"/>
         <Setter Property="Foreground" Value="white"/>
@@ -145,8 +105,8 @@
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="Button">
-                    <Border Height="30"
-                            CornerRadius="15"
+                    <Border Height="25"
+                            CornerRadius="5"
                             Background="{TemplateBinding Background}">
                         <ContentPresenter VerticalAlignment="Center"
                                           HorizontalAlignment="Center"/>
@@ -164,7 +124,6 @@
     <Style x:Key="btnDetail" TargetType="Button">
         <Setter Property="Background" Value="{StaticResource detailNomal}"/>
         <Setter Property="FontSize" Value="12"/>
-        <Setter Property="FontFamily" Value="Montserrat"/>
         <Setter Property="Cursor" Value="Hand"/>
         <Setter Property="BorderThickness" Value="0"/>
         <Setter Property="Foreground" Value="white"/>
@@ -173,8 +132,8 @@
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="Button">
-                    <Border Height="30"
-                            CornerRadius="15"
+                    <Border Height="25"
+                            CornerRadius="5"
                             Background="{TemplateBinding Background}">
                         <ContentPresenter VerticalAlignment="Center"
                                           HorizontalAlignment="Center"/>
@@ -192,7 +151,6 @@
     <Style x:Key="btnUpdate" TargetType="Button">
         <Setter Property="Background" Value="{StaticResource updateNomal}"/>
         <Setter Property="FontSize" Value="12"/>
-        <Setter Property="FontFamily" Value="Montserrat"/>
         <Setter Property="Cursor" Value="Hand"/>
         <Setter Property="BorderThickness" Value="0"/>
         <Setter Property="Foreground" Value="white"/>
@@ -201,8 +159,8 @@
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="Button">
-                    <Border Height="30"
-                            CornerRadius="15"
+                    <Border Height="25"
+                            CornerRadius="5"
                             Background="{TemplateBinding Background}">
                         <ContentPresenter VerticalAlignment="Center"
                                           HorizontalAlignment="Center"/>
@@ -229,8 +187,8 @@
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="Button">
-                    <Border Height="30"
-                            CornerRadius="15"
+                    <Border Height="25"
+                            CornerRadius="5"
                             Background="{TemplateBinding Background}">
                         <ContentPresenter VerticalAlignment="Center"
                                           HorizontalAlignment="Center"/>
@@ -257,8 +215,8 @@
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="Button">
-                    <Border Height="30"
-                            CornerRadius="15"
+                    <Border Height="25"
+                            CornerRadius="5"
                             Background="{TemplateBinding Background}">
                         <ContentPresenter VerticalAlignment="Center"
                                           HorizontalAlignment="Center"/>
@@ -273,14 +231,12 @@
         </Style.Triggers>
     </Style>
 
-    <!--Cruds button icon-->
     <Style x:Key = "crudsButtonIcon" TargetType = "fa:IconImage" >
         <Setter Property = "Foreground" Value = "Black" />
         <Setter Property = "Width" Value = "14" />
         <Setter Property = "Height" Value = "14" />
     </Style>
 
-    <!--Cruds button text-->
     <Style x:Key = "crudsButtonText" TargetType = "TextBlock" >
         <Setter Property = "Foreground" Value = "White" />
         <Setter Property = "FontFamily" Value = "Montserrat" />
