+++ ThesisManagement/Resources/Styles/Styles.xaml
@@ -3,6 +3,10 @@
 
     <Style TargetType="Window" x:Key="windowStyle">
         <Setter Property="FontFamily" Value="Montserrat" />
+        <Setter Property="FontSize" Value="14" />
+        <Setter Property="ResizeMode" Value="CanResize" />
+        <Setter Property="AllowsTransparency" Value="True" />
+        <Setter Property="WindowStyle" Value="None" />
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="Window" >
