+++ ThesisManagement/ThesisManagement.csproj
@@ -18,7 +18,6 @@
 
   <ItemGroup>
     <Folder Include="Models\" />
-    <Folder Include="CustomControls\" />
     <Folder Include="Repositories\" />
   </ItemGroup>
