+++ ThesisManagement/Views/Admin/AdminMainView.xaml
@@ -6,11 +6,10 @@
         xmlns:local="clr-namespace:ThesisManagement.Views.Admin"
         xmlns:fa="http://schemas.awesome.incremented/wpf/xaml/fontawesome.sharp"
         xmlns:viewModel="clr-namespace:ThesisManagement.ViewModels"
+        xmlns:customControl="clr-namespace:ThesisManagement.CustomControls"
         mc:Ignorable="d"
         Title="AdminMainView" Height="600" Width="1100"
         WindowStartupLocation="CenterScreen"
-        WindowStyle="None"
-        AllowsTransparency="True"
         Style="{StaticResource windowStyle}">
 
     <Window.DataContext>
@@ -70,27 +69,7 @@
                         <RowDefinition Height="*" />
                     </Grid.RowDefinitions>
 
-                    <StackPanel Orientation="Horizontal"
-                                FlowDirection="RightToLeft"
-                                MouseEnter="pnlControlBar_MouseEnter">
-                        <Button x:Name="btnClose"
-                                Click="btnClose_Click"
-                                Style="{StaticResource windowControlBtn}">
-                            <fa:IconImage Icon="X" Style="{StaticResource windowControlIcon}" />
-                        </Button>
-
-                        <Button x:Name="btnMaximize"
-                                Click="btnMaximize_Click"
-                                Style="{StaticResource windowControlBtn}">
-                            <fa:IconImage Icon="Square" Style="{StaticResource windowControlIcon}" />
-                        </Button>
-
-                        <Button x:Name="btnMinimize"
-                                Click="btnMinimize_Click"
-                                Style="{StaticResource windowControlBtn}">
-                            <fa:IconImage Icon="WindowMinimize" Style="{StaticResource windowControlIcon}" />
-                        </Button>
-                    </StackPanel>
+                    <customControl:WindowControlBar />
 
                     <Grid Grid.Row="2">
                         <ContentControl Content="{Binding CurrentChildView}"
