+++ ThesisManagement/Views/Professor/ProfessorMainView.xaml
@@ -6,12 +6,11 @@
         xmlns:local="clr-namespace:ThesisManagement.Views.Professor"
         xmlns:fa="http://schemas.awesome.incremented/wpf/xaml/fontawesome.sharp"
         xmlns:userControls="clr-namespace:ThesisManagement.Views.Professor"
+        xmlns:customControl="clr-namespace:ThesisManagement.CustomControls"
         mc:Ignorable="d"
         Title="ProfessorMainView" Height="600" Width="1150"
         WindowStartupLocation="CenterScreen"
-        WindowStyle="None"
-        Background="Transparent"
-        AllowsTransparency="True">
+        Style="{StaticResource windowStyle}">
 
     <Border CornerRadius="0">
 
@@ -57,46 +56,18 @@
             </Border>
 
             <!--Main Section-->
-            <Border CornerRadius="0" Grid.Column="1">
-                <Border.Background>
-                    <SolidColorBrush Color="{StaticResource secondary}" />
-                </Border.Background>
-
+            <Border Grid.Column="1">
                 <Grid>
                     <Grid.RowDefinitions>
                         <!--Control Bars-->
-                        <RowDefinition Height="35"/>
+                        <RowDefinition Height="25"/>
                         <!--Header / Caption-->
                         <RowDefinition Height="45"/>
                         <!--Content Section -->
                         <RowDefinition Height="*"/>
                     </Grid.RowDefinitions>
 
-                    <!--Control Bar-->
-                    <StackPanel x:Name="pnlControlBar"
-                            Grid.Row="0"
-                            Orientation="Horizontal"
-                            FlowDirection="RightToLeft"
-                            Margin="0,0,5,0">
-
-                        <Button x:Name="btnClose"
-                                Style="{StaticResource controlButton}"
-                                Tag="{StaticResource color1}">
-                            <fa:IconImage Icon="Xmark" Style="{StaticResource controlButtonIcon}" />
-                        </Button>
-
-                        <Button x:Name="btnMaximize"
-                                Style="{StaticResource controlButton}"
-                                Tag="{StaticResource color1}">
-                            <fa:IconImage Icon="Square" Style="{StaticResource controlButtonIcon}" />
-                        </Button>
-
-                        <Button x:Name="btnMinimize"
-                                Style="{StaticResource controlButton}"
-                                Tag="{StaticResource color1}">
-                            <fa:IconImage Icon="WindowMinimize" Style="{StaticResource controlButtonIcon}" />
-                        </Button>
-                    </StackPanel>
+                    <customControl:WindowControlBar />
 
                     <!--Header / Caption-->
                     <Grid Grid.Row="1">
@@ -112,11 +83,9 @@
                             <fa:IconImage Icon="RectangleList"
                                       Width="25"
                                       Height="25"
-                                      Foreground="{StaticResource titleColor1}"
                                       Margin="35,0,10,0"/>
                             <TextBlock Text="Đề tài"
-                                       Foreground="{StaticResource titleColor1}"
-                                       FontSize="20"
+                                       FontSize="17"
                                        FontFamily="Montserrat"
                                        FontWeight="Medium"
                                        VerticalAlignment="Center"/>
@@ -126,16 +95,13 @@
                                 Orientation="Horizontal"
                                 FlowDirection="RightToLeft"
                                 Margin="0,0,10,0">
-                            <Button Style="{StaticResource iconButton}"
+                            <Button Style="{StaticResource profileBtn}"
                                 Tag="{StaticResource color1}">
-                                <fa:IconImage Icon="AngleDown" Style="{StaticResource iconButtonIcon}"/>
+                                <fa:IconImage Icon="AngleDown" Style="{StaticResource profileBtnIcon}"/>
                             </Button>
 
                             <TextBlock Text="Nguyễn Thủy An" 
-                                       FontSize="20"
-                                       Foreground="{StaticResource titleColor1}"
                                        FontWeight="Medium"
-                                       FontFamily="Montserrat"
                                        VerticalAlignment="Center"/>
 
                         </StackPanel>
